pants: fix errorformat to include column number
This change fixes the error format used to detect errors within a pants
project. The scala compiler includes the column number as well in the
output and that needs to be reflected within the error format. Fixing
this allows the correct file to be opened after compilation so that the
developer can make a fix. Note: I believe this only applies to scala
compilation.